fix(audit): address CodeRabbit review; split audit_report to hold self-scan gate
Review fixes:
- is_finding_id operates on bytes (no UTF-8 boundary panic on untrusted ids)
- validate() is registry-authoritative: exact department membership,
  enabled/status consistency, score+coverage entries must reference a
  department run in the report
- escape markdown table cells in rendered audit sections
- docs/rubrics document lowercase wire values; registry slot test asserts
  the three disabled department slots

Gate fix:
- split audit_report.rs (1721 nonblank lines, god-file) into
  src/audit_report/{enums,json_helpers,model,registry,validate,render}
  plus sibling test files; sentrux gate now reports God 29->29,
  Coupling 44.78->44.59, Quality 3971->3973, Cycles 0->0
- update test-harness #[path] includes and internalization digests

Refs #18
