+++ crates/code-intel-cli/src/audit_report/mod.rs
@@ -1,0 +1,12 @@
+#[path = "../content_contract.rs"]
+mod content_contract;
+
+mod enums;
+mod json_helpers;
+mod model;
+mod registry;
+mod render;
+mod validate;
+
+pub(crate) use model::AuditReport;
+pub(crate) use render::render_markdown_section;
