+++ crates/code-intel-cli/tests/artifact_index.rs
@@ -12,7 +12,7 @@ mod adapter_contract;
 mod artifact_index;
 #[path = "../src/artifact_ref.rs"]
 mod artifact_ref;
-#[path = "../src/audit_report.rs"]
+#[path = "../src/audit_report/mod.rs"]
 mod audit_report;
 #[path = "../src/capability.rs"]
 mod capability;
