+++ crates/code-intel-cli/tests/audit_report.rs
@@ -152,6 +152,23 @@ fn registry_file_has_unique_department_ids_and_existing_rubric_and_finding_contr
         root.join(finding_contract).is_file(),
         "findingContract file must exist on disk: {finding_contract}"
     );
+    // T1 registers exactly these three departments and does not yet run any
+    // of them (see docs/audit-report.md and CHANGELOG.md). Check each
+    // required slot by id rather than the full list, so a future PR can add
+    // more departments without this test needing to change.
+    for id in ["security", "ai-safety", "supply-chain"] {
+        let department = registry["departments"]
+            .as_array()
+            .unwrap()
+            .iter()
+            .find(|department| department["id"] == id)
+            .unwrap_or_else(|| panic!("registry must contain a department with id \"{id}\""));
+        assert_eq!(
+            department["enabled"],
+            json!(false),
+            "department \"{id}\" must currently be enabled: false"
+        );
+    }
 }
 
 #[test]
