+++ crates/code-intel-cli/tests/run_commit.rs
@@ -9,7 +9,7 @@ use serde_json::{json, Value};
 mod adapter_contract;
 #[path = "../src/artifact_ref.rs"]
 mod artifact_ref;
-#[path = "../src/audit_report.rs"]
+#[path = "../src/audit_report/mod.rs"]
 mod audit_report;
 #[path = "../src/capability.rs"]
 mod capability;
