+++ crates/code-intel-cli/tests/sentrux_adapter.rs
@@ -12,7 +12,7 @@ mod adapter_contract;
 mod admissibility;
 #[path = "../src/artifact_ref.rs"]
 mod artifact_ref;
-#[path = "../src/audit_report.rs"]
+#[path = "../src/audit_report/mod.rs"]
 mod audit_report;
 #[path = "../src/capability.rs"]
 mod capability;
