+++ docs/audit-report.md
@@ -19,9 +19,9 @@ Every finding in `audit-report.json.findings` is one object with these fields. A
 | `id` | string, pattern `^[a-z0-9-]+-[0-9]{3}$` | yes | Stable id, e.g. `security-001`. |
 | `department` | string | yes | Registered department id from `orchestration/audit/departments.v1.json`. |
 | `title` | string | yes | Short summary. |
-| `severity` | `Critical` / `High` / `Medium` / `Low` / `Info` | yes | See `orchestration/audit/rubrics/severity.md`. |
-| `confidence` | `High` / `Medium` / `Low` | yes | See `orchestration/audit/rubrics/confidence.md`. |
-| `status` | `Confirmed` / `Suspected` | yes | Whether the evidence directly confirms the finding. |
+| `severity` | `critical` / `high` / `medium` / `low` / `info` | yes | See `orchestration/audit/rubrics/severity.md`. |
+| `confidence` | `high` / `medium` / `low` | yes | See `orchestration/audit/rubrics/confidence.md`. |
+| `status` | `confirmed` / `suspected` | yes | Whether the evidence directly confirms the finding. |
 | `affected_area` | string | no | Module, component, or subsystem. |
 | `evidence` | array of evidence refs, >= 1 entry | yes | See `orchestration/audit/rubrics/evidence.md`; each entry carries `kind` (`file` / `modality_signal` / `command` / `manual_read`), `source`, and optionally `path`, `line_start`, `line_end`, `modality`, `note`. |
 | `problem` | string | yes | One-paragraph description of the issue. |
@@ -41,11 +41,13 @@ Findings must never write secret material in plaintext. A finding about a leaked
 `crates/code-intel-cli/src/audit_report.rs` parses and validates every `audit-report.json`. Parsing itself enforces the JSON Schema contract (required fields, closed objects — no `additionalProperties`, enum values, the finding `id` pattern, the `evidence` minItems). `validate()` then enforces invariants the schema cannot express, each producing a distinct error:
 
 1. Every finding has at least one evidence entry (schema-enforced), and a finding with `status: confirmed` has at least one `file`-kind evidence entry with a `path`.
-2. Every `finding.department`, `score_dashboard` entry department, and `coverage_matrix` row department is a department id registered in `orchestration/audit/departments.v1.json`.
-3. A department whose `status` is `not_assessed` or `disabled` has a null score in `score_dashboard` and a `not_assessed` row in `coverage_matrix`.
-4. `score_dashboard.overall` is the mean of the non-null entry scores, rounded to 1 decimal place, or `null` when nothing scored; `validate()` recomputes it and rejects a mismatch.
-5. A department that scores `10.0` with zero findings must have `coverage: high` — see `orchestration/audit/rubrics/scoring.md`.
-6. Every department listed in the report has exactly one `coverage_matrix` row and at most one `score_dashboard` entry.
+2. Every `finding.department`, `score_dashboard` entry department, and `coverage_matrix` row department is a department id registered in `orchestration/audit/departments.v1.json`; a `score_dashboard` entry's and a `coverage_matrix` row's department must additionally be one of the department ids actually listed in the report's own `departments` array, not merely registered — a registered id with no corresponding department run cannot carry a score or coverage entry.
+3. The report's `departments` array exactly matches the registry: every registered department has a run in the report, and every department run in the report names a registered department.
+4. A department run's `status` is consistent with the registry's `enabled` flag for that department: `enabled: false` requires `status: disabled`, and `enabled: true` requires `status: assessed` or `not_assessed` (never `disabled`).
+5. A department whose `status` is `not_assessed` or `disabled` has a null score in `score_dashboard` and a `not_assessed` row in `coverage_matrix`.
+6. `score_dashboard.overall` is the mean of the non-null entry scores, rounded to 1 decimal place, or `null` when nothing scored; `validate()` recomputes it and rejects a mismatch.
+7. A department that scores `10.0` with zero findings must have `coverage: high` — see `orchestration/audit/rubrics/scoring.md`.
+8. Every department listed in the report has exactly one `coverage_matrix` row and at most one `score_dashboard` entry.
 
 The registry itself (`orchestration/audit/departments.v1.json`) has its own invariants, checked by `DepartmentRegistry::validate()`: department ids are unique, every rubric file it points at exists on disk, and every `enabled: true` department's prompt file exists on disk. A disabled department may point at a prompt file that does not exist yet — that file is the department ticket's job, not the kernel's.
